requirements.txt
Flask~=2.0.1
APScheduler~=3.7.0
blaseball-mike~=6.4.1
lark-parser~=0.11.3